refactoring - updated tests
