+++ bfxapi/examples/rest/create_order.py
@@ -3,7 +3,8 @@ import sys
 import asyncio
 import time
 sys.path.append('../../../')
-from bfxapi import Client, WS_HOST, REST_HOST
+from bfxapi import Client
+from bfxapi.constants import WS_HOST, REST_HOST
 from bfxapi.models import OrderType
 
 API_KEY=os.getenv("BFX_KEY")
