+++ bfxapi/examples/rest/get_authenticated_data.py
@@ -4,7 +4,8 @@ import asyncio
 import time
 sys.path.append('../../../')
 
-from bfxapi import Client, WS_HOST, REST_HOST
+from bfxapi import Client
+from bfxapi.constants import WS_HOST, REST_HOST
 
 API_KEY=os.getenv("BFX_KEY")
 API_SECRET=os.getenv("BFX_SECRET")
