+++ bfxapi/examples/rest/get_seed_trades.py
@@ -3,7 +3,8 @@ import sys
 import asyncio
 sys.path.append('../../../')
 
-from bfxapi import Client, PUB_WS_HOST, PUB_REST_HOST
+from bfxapi import Client
+from bfxapi.constants import PUB_WS_HOST, PUB_REST_HOST
 
 # Retrieving seed trades requires public hosts
 bfx = Client(
