+++ bfxapi/examples/ws/cancel_order.py
@@ -2,7 +2,8 @@ import os
 import sys
 sys.path.append('../../../')
 
-from bfxapi import Client, Order, WS_HOST, REST_HOST
+from bfxapi import Client, Order
+from bfxapi.constants import WS_HOST, REST_HOST
 
 API_KEY=os.getenv("BFX_KEY")
 API_SECRET=os.getenv("BFX_SECRET")
