+++ bfxapi/examples/ws/resubscribe_orderbook.py
@@ -1,7 +1,8 @@
 import sys
 sys.path.append('../../../')
 
-from bfxapi import Client, PUB_WS_HOST, PUB_REST_HOST
+from bfxapi import Client
+from bfxapi.constants import PUB_WS_HOST, PUB_REST_HOST
 
 # Retrieving orderbook requires public hosts
 bfx = Client(
