+++ bfxapi/examples/ws/subscribe_orderbook.py
@@ -2,7 +2,8 @@ import os
 import sys
 sys.path.append('../../../')
 
-from bfxapi import Client, PUB_WS_HOST, PUB_REST_HOST
+from bfxapi import Client
+from bfxapi.constants import PUB_WS_HOST, PUB_REST_HOST
 
 # Retrieving trades/candles requires public hosts
 bfx = Client(
