+++ bfxapi/examples/ws/wallet_balance.py
@@ -1,7 +1,8 @@
 import os
 import sys
 sys.path.append('../../../')
-from bfxapi import Client, WS_HOST, REST_HOST
+from bfxapi import Client
+from bfxapi.constants import WS_HOST, REST_HOST
 
 API_KEY=os.getenv("BFX_KEY")
 API_SECRET=os.getenv("BFX_SECRET")
